+++ Core/Core.csproj
@@ -191,6 +191,7 @@
     <Compile Include="Menu\IGMData\Base.cs" />
     <Compile Include="Menu\IGMData\Commands.cs" />
     <Compile Include="Menu\IGMData\Dialog\Confirm.cs" />
+    <Compile Include="Menu\IGMData\Group\Base.cs" />
     <Compile Include="Menu\IGMData\Pool\Draw.cs" />
     <Compile Include="Menu\IGMData\Pool\Draw\Commands.cs" />
     <Compile Include="Menu\IGMData\Pool\BlueMagic.cs" />
@@ -783,7 +784,6 @@
     <Compile Include="Menu\Images\Icons.ID.cs" />
     <Compile Include="Menu\Images\Icons.InsertCustomEntries.cs" />
     <Compile Include="Menu\IGMData\Obsolete\IGMData_Container.cs" />
-    <Compile Include="Menu\IGMData\IGMData_Group.cs" />
     <Compile Include="Menu\IGMData\IGMData_Pool.cs" />
     <Compile Include="Menu\IGM\IGM.cs" />
     <Compile Include="Menu\IGM\IGMData\IGMData_Clock.cs" />
