+++ Core/Menu/IGMData/Group/Base.cs
@@ -1,14 +1,14 @@
 ﻿using System.Diagnostics;
 
-namespace OpenVIII
+namespace OpenVIII.IGMData.Group
 {
-    public class IGMData_Group : IGMData.Base
+    public class Base : IGMData.Base
     {
         #region Constructors
 
-        public IGMData_Group(params IGMData.Base[] d) : base(d.Length, 1, container: new IGMDataItem.Empty()) => Init(d);
+        public Base(params IGMData.Base[] d) : base(d.Length, 1, container: new IGMDataItem.Empty()) => Init(d);
 
-        public IGMData_Group() : base(container: new IGMDataItem.Empty()) => Debug.WriteLine($"{this} :: Not init may need to call it later");
+        public Base() : base(container: new IGMDataItem.Empty()) => Debug.WriteLine($"{this} :: Not init may need to call it later");
 
         protected virtual void Init(IGMData.Base[] d, bool baseinit = false)
         {
